+++ src/shapes/Object/InteractiveObject.ts
@@ -284,14 +284,16 @@ export class InteractiveFabricObject<
       this.cornerSize,
       position.x,
       position.y,
-      false
+      false,
+      this
     );
     const touchCorner = control.calcCornerCoords(
       angle,
       this.touchCornerSize,
       position.x,
       position.y,
-      true
+      true,
+      this
     );
     return { corner, touchCorner };
   }